requirements_dev.txt
pytest-xdist
stdeb
debhelper
build
mypy
pytest-django
